+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 0.13.0
 + Add `BalanceNotEnoughException`
 + Refactor `BalanceNotEnoughException`
 + Refactor CORS